Rename the RHF directory to HF.
