+++ gqcp/include/QCMethod/CMakeLists.txt
@@ -8,6 +8,6 @@ target_sources(gqcp
 add_subdirectory(Applications)
 add_subdirectory(CI)
 add_subdirectory(Geminals)
+add_subdirectory(HF)
 add_subdirectory(OrbitalOptimization)
-add_subdirectory(RHF)
 add_subdirectory(RMP2)
