+++ gqcp/src/QCMethod/Applications/DOCIRHF.cpp
@@ -21,7 +21,7 @@
 #include "Basis/SpinorBasis/RSpinorBasis.hpp"
 #include "QCMethod/CI/CISolver.hpp"
 #include "Operator/SecondQuantized/SQHamiltonian.hpp"
-#include "QCMethod/RHF/DIISRHFSCFSolver.hpp"
+#include "QCMethod/HF/DIISRHFSCFSolver.hpp"
 
 
 namespace GQCP {
