+++ gqcp/tests/QCMethod/Applications/constrained_RHF_test.cpp
@@ -22,7 +22,7 @@
 #include "Basis/transform.hpp"
 #include "Operator/SecondQuantized/SQHamiltonian.hpp"
 #include "Processing/Properties/expectation_values.hpp"
-#include "QCMethod/RHF/DIISRHFSCFSolver.hpp"
+#include "QCMethod/HF/DIISRHFSCFSolver.hpp"
 #include "QCModel/RHF/RHF.hpp"
 
 #include <random>
