+++ gqcp/tests/QCMethod/CI/CISolver_Hubbard_Dense_test.cpp
@@ -24,7 +24,7 @@
 #include "QCMethod/CI/CISolver.hpp"
 #include "QCMethod/CI/HamiltonianBuilder/FCI.hpp"
 #include "QCMethod/CI/HamiltonianBuilder/Hubbard.hpp"
-#include "QCMethod/RHF/PlainRHFSCFSolver.hpp"
+#include "QCMethod/HF/PlainRHFSCFSolver.hpp"
 
 
 BOOST_AUTO_TEST_CASE ( test_Hubbard_vs_FCI_dense ) {
