+++ gqcp/tests/QCMethod/CI/HamiltonianBuilder/DOCI_test.cpp
@@ -22,7 +22,7 @@
 #include "QCMethod/CI/HamiltonianBuilder/DOCI.hpp"
 
 #include "Operator/SecondQuantized/SQHamiltonian.hpp"
-#include "QCMethod/RHF/PlainRHFSCFSolver.hpp"
+#include "QCMethod/HF/PlainRHFSCFSolver.hpp"
 
 
 BOOST_AUTO_TEST_CASE ( DOCI_constructor ) {
