+++ gqcp/tests/QCMethod/CMakeLists.txt
@@ -1,8 +1,8 @@
 add_subdirectory(Applications)
 add_subdirectory(CI)
 add_subdirectory(Geminals)
+add_subdirectory(HF)
 add_subdirectory(OrbitalOptimization)
-add_subdirectory(RHF)
 add_subdirectory(RMP2)
 
 set(test_target_sources ${test_target_sources} PARENT_SCOPE)
